First fully working version
eTSDB updated with unlimited number of columns.
SDI12 fully implemented.
Virtual machine fully working.
Log is now capable to change FD.
Dynamic vector is added to mem structs.

+++ Inc/SPIFlash.h
@@ -67,9 +67,9 @@ namespace FwLogger
 			uint16_t rx_idx;
 			uint16_t rx_size;
 
-			uint16_t _last_us; // para añadir un poco de espera por si hace falta para que a la memoria le de tiempo de hacer cosas
+			uint32_t _last_us; // para añadir un poco de espera por si hace falta para que a la memoria le de tiempo de hacer cosas
 			uint32_t _last_write_ms;
-			uint16_t _max_delay_time;
+			uint32_t _max_delay_time;
 
 			int writeEnable(); // añadir bsy check
 			int writeDisable();
